Split workspace cleanup FSM from workspace operations
